+++ src/data_sources/mongo/services/mod.rs
@@ -1,4 +1,4 @@
-use bson::Document;
+use bson::{doc, Document};
 use log::debug;
 
 // mod create_many;
@@ -30,5 +30,37 @@ impl Services {
             }
         }
         set_doc
+    }
+    pub fn create_nested_find_filter(doc: &Document) -> Document {
+        debug!("Creating Nested Find Filter");
+        debug!("Initial Document/Filter: {:?}", doc);
+        let mut find_doc = Document::new();
+        for (key, value) in doc.clone().iter_mut() {
+            if let Some(sub_doc) = value.as_document() {
+                debug!("Found Sub Document: {:?}", sub_doc);
+                let sub_set_doc = Services::create_nested_find_filter(sub_doc);
+                for (sub_key, sub_value) in sub_set_doc.iter() {
+                    let nested_key = format!("{}.{}", key, sub_key);
+                    find_doc.insert(nested_key, sub_value.clone());
+                }
+            } else {
+                if let Some(array) = value.as_array() {
+                    let is_docs = array.iter().all(|bson| bson.as_document().is_some());
+                    if is_docs {
+                        let mut docs = vec![];
+                        for b in array {
+                            docs.push(doc! { key.clone(): {"$elemMatch": b}})
+                        }
+                        find_doc.insert("$and", docs);
+                    } else {
+                        find_doc.insert(key.clone(), doc! { "$in": array });
+                    }
+                    continue;
+                }
+                find_doc.insert(key.clone(), value.clone());
+            }
+        }
+        debug!("Created Nested Find Filter: {:?}", find_doc);
+        find_doc
     }
 }